Moved KeysAndCertsGenerator and DetRandomProvider to consensus-fakes
