+++ platform-sdk/base-crypto/src/test/java/org/hiero/base/crypto/SerializablePublicKeyTests.java
@@ -10,6 +10,7 @@ import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
 import java.util.stream.Stream;
 import org.hiero.base.utility.test.fixtures.io.InputOutputStream;
+import org.hiero.consensus.fakes.crypto.DetRandomProvider;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
