+++ platform-sdk/consensus-fakes/src/main/java/org/hiero/consensus/fakes/crypto/DetRandomProvider.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package org.hiero.base.crypto;
+package org.hiero.consensus.fakes.crypto;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.NoSuchProviderException;
@@ -20,9 +20,7 @@ public class DetRandomProvider {
      * 		if the security provider is not available on the system
      * @throws NoSuchAlgorithmException
      * 		if the algorithm is not available on the system
-     * @deprecated This class is deprecated and will be extracted to a text fixture module in a future release.
      */
-    @Deprecated(forRemoval = true)
     public static SecureRandom getDetRandom() throws NoSuchProviderException, NoSuchAlgorithmException {
         return SecureRandom.getInstance(PRNG_TYPE, PRNG_PROVIDER);
     }
