+++ platform-sdk/consensus-gossip-impl/src/test/java/org/hiero/consensus/gossip/impl/network/NetworkPeerIdentifierTest.java
@@ -27,11 +27,11 @@ import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.hiero.base.crypto.CertificateUtils;
-import org.hiero.base.crypto.DetRandomProvider;
 import org.hiero.base.crypto.KeyGeneratingException;
 import org.hiero.base.crypto.SigningSchema;
 import org.hiero.base.utility.test.fixtures.io.ResourceLoader;
 import org.hiero.consensus.crypto.KeyCertPurpose;
+import org.hiero.consensus.fakes.crypto.DetRandomProvider;
 import org.hiero.consensus.model.node.NodeId;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
