+++ platform-sdk/consensus-otter-tests/src/testIntegration/java/org/hiero/otter/fixtures/test/NetworkRosterTests.java
@@ -13,7 +13,7 @@ import java.security.cert.CertificateEncodingException;
 import java.util.List;
 import org.hiero.base.crypto.KeyGeneratingException;
 import org.hiero.base.crypto.SigningSchema;
-import org.hiero.consensus.crypto.KeysAndCertsGenerator;
+import org.hiero.consensus.fakes.crypto.KeysAndCertsGenerator;
 import org.hiero.consensus.model.node.KeysAndCerts;
 import org.hiero.otter.fixtures.Network;
 import org.hiero.otter.fixtures.Node;
