+++ platform-sdk/consensus-otter-tests/src/testOtter/java/module-info.java
@@ -8,6 +8,7 @@ open module org.hiero.otter.test {
     requires org.hiero.consensus.state;
     requires org.hiero.consensus.transaction.handling;
     requires org.hiero.otter.fixtures;
+    requires org.hiero.consensus.fakes;
     requires org.junit.jupiter.params;
     requires static com.github.spotbugs.annotations;
 }
