+++ platform-sdk/consensus-otter-tests/src/testOtter/java/org/hiero/otter/test/SigningSchemaTest.java
@@ -18,7 +18,7 @@ import java.security.SecureRandom;
 import java.time.Duration;
 import org.hiero.base.crypto.KeyGeneratingException;
 import org.hiero.base.crypto.SigningSchema;
-import org.hiero.consensus.crypto.KeysAndCertsGenerator;
+import org.hiero.consensus.fakes.crypto.KeysAndCertsGenerator;
 import org.hiero.otter.fixtures.Network;
 import org.hiero.otter.fixtures.OtterTest;
 import org.hiero.otter.fixtures.TestEnvironment;
