+++ platform-sdk/consensus-otter-tests/src/testOtter/java/org/hiero/otter/test/StartFromStateTest.java
@@ -19,11 +19,11 @@ import java.security.NoSuchProviderException;
 import java.security.SecureRandom;
 import java.time.Duration;
 import java.util.stream.Stream;
-import org.hiero.base.crypto.DetRandomProvider;
 import org.hiero.base.crypto.KeyGeneratingException;
 import org.hiero.base.crypto.SigningSchema;
-import org.hiero.consensus.crypto.KeysAndCertsGenerator;
 import org.hiero.consensus.event.creator.config.EventCreationConfig_;
+import org.hiero.consensus.fakes.crypto.DetRandomProvider;
+import org.hiero.consensus.fakes.crypto.KeysAndCertsGenerator;
 import org.hiero.consensus.state.saved.SavedStateMetadata;
 import org.hiero.otter.fixtures.Network;
 import org.hiero.otter.fixtures.Node;
