+++ platform-sdk/consensus-sloth/src/testFixtures/java/org/hiero/sloth/fixtures/internal/AbstractNetwork.java
@@ -26,7 +26,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.stream.IntStream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.hiero.consensus.crypto.KeysAndCertsGenerator;
+import org.hiero.consensus.fakes.crypto.KeysAndCertsGenerator;
 import org.hiero.consensus.model.node.KeysAndCerts;
 import org.hiero.consensus.model.node.NodeId;
 import org.hiero.consensus.model.quiescence.QuiescenceCommand;
