+++ platform-sdk/swirlds-platform-core/src/test/java/org/hiero/consensus/ConsensusLayerWiringTests.java
@@ -34,12 +34,12 @@ import java.util.function.UnaryOperator;
 import java.util.stream.Stream;
 import org.hiero.base.crypto.KeyGeneratingException;
 import org.hiero.base.utility.test.fixtures.file.TestFileSystemManager;
-import org.hiero.consensus.crypto.KeysAndCertsGenerator;
 import org.hiero.consensus.event.creator.EventCreatorModule;
 import org.hiero.consensus.event.intake.EventIntakeModule;
 import org.hiero.consensus.event.stream.ConsensusEventStream;
 import org.hiero.consensus.event.stream.config.EventConfig_;
 import org.hiero.consensus.event.stream.config.EventStreamWiringConfig;
+import org.hiero.consensus.fakes.crypto.KeysAndCertsGenerator;
 import org.hiero.consensus.fakes.noop.NoOpMetrics;
 import org.hiero.consensus.fakes.noop.NoOpRecycleBin;
 import org.hiero.consensus.gossip.GossipModule;
